requirements.txt
pandas
opencv-python
numpy
pytesseract
pdf2image
python-dotenv
google-genai
openpyxl